move start_log_file
