+++ src/odin/run.py
@@ -1,18 +1,10 @@
 import sched
 import signal
 import time
-import logging
-import threading
-from datetime import datetime
-
 
 from odin.utils.runtime import validate_env_vars
 from odin.utils.runtime import handle_sigterm
 from odin.utils.logger import ProcessLog
-from odin.utils.logger import LOGGER_NAME
-from odin.utils.logger import LOG_FORMAT
-from odin.utils.logger import DATE_FORMAT
-from odin.utils.logger import log_max_mem_usage
 from odin.migrate.process import start_migrations
 
 # Job Schedule functions
@@ -52,7 +44,7 @@ def start():
 
     schedule = sched.scheduler(time.monotonic, time.sleep)
 
-    log = ProcessLog("odin_event_loop")
+    ProcessLog("odin_event_loop")
 
     # Schedule ODIN Jobs
     schedule_sigterm_check(schedule)
@@ -60,29 +52,3 @@ def start():
     schedule_cubic_ods_fact_gen(schedule)
 
     schedule.run()
-    log.complete()
-
-
-def start_log_file():
-    """Application Entry for logging to local file instead of stderr/stdout."""
-    logger = logging.getLogger(LOGGER_NAME)
-    for handler in logger.handlers:
-        logger.removeHandler(handler)
-
-    now = datetime.now().replace(microsecond=0)
-    log_file_name = f"./odin_log_{now.isoformat()}.log"
-    handler = logging.FileHandler(log_file_name, mode="w")
-    handler.setFormatter(logging.Formatter(fmt=LOG_FORMAT, datefmt=DATE_FORMAT, style="{"))
-    logger.addHandler(handler)
-
-    stop_event = threading.Event()
-    mem_thread = threading.Thread(target=log_max_mem_usage, args=(stop_event,))
-    mem_thread.start()
-
-    try:
-        start()
-    except Exception as exception:
-        raise exception
-    finally:
-        stop_event.set()
-        mem_thread.join()
